refactor: data dictionary table options (#497)
