+++ src/components/DataDictionary/components/Table/options/core.ts
@@ -1,6 +1,6 @@
 import { CoreOptions, getCoreRowModel, RowData } from "@tanstack/react-table";
-import { ROW_POSITION } from "../../../../../Table/features/RowPosition/constants";
-import { ROW_PREVIEW } from "../../../../../Table/features/RowPreview/constants";
+import { ROW_POSITION } from "../../../../Table/features/RowPosition/constants";
+import { ROW_PREVIEW } from "../../../../Table/features/RowPreview/constants";
 
 export const CORE_OPTIONS: Pick<
   CoreOptions<RowData>,
