+++ src/components/DataDictionary/components/Table/options/expanded.ts
